Fix regression (filteredTorrents)
